Split LocalTokenizerService into its own file
So it can be used directly, regardless of the environment variable
