+++ util/tokenizer_service.js
@@ -13,50 +13,9 @@ const Tp = require('thingpedia');
 
 const net = require('net');
 const JsonDatagramSocket = require('./json_datagram_socket');
+const LocalTokenizerService = require('./local_tokenizer_service');
 
 const Config = require('../config');
-
-class LocalTokenizerService {
-    constructor() {
-        const socket = new net.Socket();
-        socket.connect({
-            host: '127.0.0.1',
-            port: 8888
-        });
-        this._socket = new JsonDatagramSocket(socket, socket, 'utf8');
-        this._socket.on('data', (msg) => {
-            let req = this._requests.get(msg.req);
-            if (!req)
-                return;
-            this._requests.delete(msg.req);
-
-            if (msg.error) {
-                req.reject(new Error(msg.error));
-            } else {
-                req.resolve({
-                    tokens: msg.tokens,
-                    entities: msg.values
-                });
-            }
-        });
-
-        this._requests = new Map();
-        this._nextRequest = 0;
-    }
-
-    end() {
-        this._socket.end();
-    }
-
-    tokenize(languageTag, utterance) {
-        const reqId = this._nextRequest++;
-        return new Promise((resolve, reject) => {
-            this._requests.set(reqId, { resolve, reject });
-
-            this._socket.write({ req: reqId, utterance, languageTag });
-        });
-    }
-}
 
 function tokenizeWithAlmondNL(language, sentence) {
     let url = Config.NL_SERVER_URL + '/' + language + '/tokenize?q=' + encodeURIComponent(sentence);
